chore: add Team and Dashboard items
